# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: AI Schedule Manager API
  version: 1.0.0
  description: |
    Neural-powered scheduling API for small businesses with complete CRUD operations,
    natural language rule parsing, and AI-driven schedule optimization.

    ## Features
    - Complete employee, schedule, and rule management
    - Natural language rule parsing using NLP
    - AI-powered schedule optimization
    - Real-time notifications
    - Role-based access control
    - Comprehensive analytics

    ## Authentication
    This API uses JWT tokens for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

  contact:
    name: AI Schedule Manager API Support
    email: support@ai-schedule-manager.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.ai-schedule-manager.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /api/auth/login:
    post:
      summary: User Login
      description: Authenticate user and return JWT tokens
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: "manager@example.com"
              password: "securepassword123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/employees:
    get:
      summary: List Employees
      description: Get all employees with pagination and filtering
      tags: [Employees]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
        - name: role
          in: query
          description: Filter by employee role
          schema:
            $ref: '#/components/schemas/EmployeeRole'
        - name: active
          in: query
          description: Filter by active status
          schema:
            type: boolean
        - name: search
          in: query
          description: Search by name or email
          schema:
            type: string
      responses:
        '200':
          description: List of employees
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/EmployeeResponse'

    post:
      summary: Create Employee
      description: Create a new employee (manager only)
      tags: [Employees]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeCreate'
      responses:
        '201':
          description: Employee created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/employees/{employee_id}:
    get:
      summary: Get Employee
      description: Get specific employee by ID
      tags: [Employees]
      parameters:
        - $ref: '#/components/parameters/EmployeeIdParam'
      responses:
        '200':
          description: Employee details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      summary: Update Employee
      description: Update employee information (manager only)
      tags: [Employees]
      parameters:
        - $ref: '#/components/parameters/EmployeeIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeUpdate'
      responses:
        '200':
          description: Employee updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      summary: Delete Employee
      description: Delete employee (manager only)
      tags: [Employees]
      parameters:
        - $ref: '#/components/parameters/EmployeeIdParam'
      responses:
        '200':
          description: Employee deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Employee deleted successfully"
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/employees/{employee_id}/schedule:
    get:
      summary: Get Employee Schedule
      description: Get schedule for specific employee
      tags: [Employees, Schedules]
      parameters:
        - $ref: '#/components/parameters/EmployeeIdParam'
        - name: date_from
          in: query
          description: Start date filter
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          description: End date filter
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Employee schedule
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScheduleResponse'

  # Rules
  /api/rules:
    get:
      summary: List Rules
      description: Get all rules with pagination and filtering
      tags: [Rules]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
        - name: rule_type
          in: query
          description: Filter by rule type
          schema:
            $ref: '#/components/schemas/RuleType'
        - name: employee_id
          in: query
          description: Filter by employee ID
          schema:
            type: integer
        - name: active
          in: query
          description: Filter by active status
          schema:
            type: boolean
      responses:
        '200':
          description: List of rules
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/RuleResponse'
  /api/rules/parse:
    post:
      summary: Parse Natural Language Rule
      description: Parse natural language text into structured rule
      tags: [Rules]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleParseRequest'
            example:
              rule_text: "John cannot work on Sundays and prefers morning shifts"
      responses:
        '200':
          description: Rule parsed and created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/rules/{rule_id}:
    get:
      summary: Get Rule
      description: Get specific rule by ID
      tags: [Rules]
      parameters:
        - $ref: '#/components/parameters/RuleIdParam'
      responses:
        '200':
          description: Rule details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      summary: Update Rule
      description: Update rule (manager only)
      tags: [Rules]
      parameters:
        - $ref: '#/components/parameters/RuleIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleUpdate'
      responses:
        '200':
          description: Rule updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      summary: Delete Rule
      description: Delete rule (manager only)
      tags: [Rules]
      parameters:
        - $ref: '#/components/parameters/RuleIdParam'
      responses:
        '200':
          description: Rule deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Rule deleted successfully"
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'

  # Schedules
  /api/schedules:
    get:
      summary: List Schedules
      description: Get all schedules with pagination and filtering
      tags: [Schedules]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
        - name: employee_id
          in: query
          description: Filter by employee ID
          schema:
            type: integer
        - name: shift_id
          in: query
          description: Filter by shift ID
          schema:
            type: integer
        - name: date_from
          in: query
          description: Start date filter
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          description: End date filter
          schema:
            type: string
            format: date
        - name: status
          in: query
          description: Filter by schedule status
          schema:
            $ref: '#/components/schemas/ScheduleStatus'
      responses:
        '200':
          description: List of schedules
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/ScheduleResponse'
  /api/schedules/{schedule_id}:
    get:
      summary: Get Schedule
      description: Get specific schedule by ID
      tags: [Schedules]
      parameters:
        - $ref: '#/components/parameters/ScheduleIdParam'
      responses:
        '200':
          description: Schedule details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      summary: Update Schedule
      description: Update schedule (manager only)
      tags: [Schedules]
      parameters:
        - $ref: '#/components/parameters/ScheduleIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleUpdate'
      responses:
        '200':
          description: Schedule updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      summary: Delete Schedule
      description: Delete schedule (manager only)
      tags: [Schedules]
      parameters:
        - $ref: '#/components/parameters/ScheduleIdParam'
      responses:
        '200':
          description: Schedule deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Schedule deleted successfully"
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'

  # Schedule Generation and Optimization
  /api/schedule/generate:
    post:
      summary: Generate Schedule
      description: Generate optimized schedule for date range using AI
      tags: [AI Scheduling]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleGenerateRequest'
            example:
              start_date: "2024-01-01"
              end_date: "2024-01-07"
              constraints:
                max_hours_per_employee: 40
                min_coverage_ratio: 1.2
      responses:
        '200':
          description: Schedule generation started
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1234
                  start_date:
                    type: string
                    format: date
                  end_date:
                    type: string
                    format: date
                  status:
                    type: string
                    example: "generated"
                  shifts:
                    type: array
                    items:
                      type: object
                  created_at:
                    type: string
                    format: date-time
                  message:
                    type: string
                    example: "Schedule generation started. Check back for results."
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/notifications/{notification_id}/read:
    patch:
      summary: Mark Notification as Read
      description: Mark specific notification as read
      tags: [Notifications]
      parameters:
        - $ref: '#/components/parameters/NotificationIdParam'
      responses:
        '200':
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/analytics/overview:
    get:
      summary: Get Analytics Overview
      description: Get comprehensive analytics overview (manager only)
      tags: [Analytics]
      responses:
        '200':
          description: Analytics overview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsOverview'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (starts from 1)
      schema:
        type: integer
        minimum: 1
        default: 1
...
  schemas:
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "manager@example.com"
        password:
          type: string
          minLength: 1
          example: "securepassword123"
...